Separate vendor parsing into distinct modules
